cmake: Set the same soname as Makefile.am (#168)
In Makefile.am, soname is set to 1:0:0, but in CMake it is set to 0:0:0.
This means that when building with CMake, libraries with different sonames will
be built. This will not maintain library compatibility.
This sets the same soname in cmake as well.

Before:
```
$ mkdir t && cd t && cmake .. && make
$ ls *.so*
libbzip3.so  libbzip3.so.0  libbzip3.so.0.0.0
$ objdump -p libbzip3.so
  SONAME               libbzip3.so.0.0.0
$ cd .. && autoreconf -ivf && ./configure && make
$ ls .libs/*.so*
.libs/libbzip3.so .libs/libbzip3.so.1 .libs/libbzip3.so.1.0.0
$ objdump -p ./.libs/libbzip3.so | grep SONAME
  SONAME               libbzip3.so.1
```

After:
```
$ mkdir t && cd t && cmake .. && make
$ ls *.so*
libbzip3.so  libbzip3.so.1  libbzip3.so.1.0.0
$ objdump -p libbzip3.so
  SONAME               libbzip3.so.1
```

Signed-off-by: Nobuhiro Iwamatsu <iwamatsu@nigauri.org>